requirements.txt
influxdb-client

minimalmodbus

paho-mqtt<2.0.0
 
